pyproject.toml
[project]
name = "ensemble"
version = "0.1.0"
description = "LLM Prediction Market Ensemble - Research Framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.6",
    "pandas>=2.0",
    "httpx>=0.27",
    "typer>=0.12",
    "rich>=13.0",
    "python-dateutil>=2.9",
    "openai>=1.60",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.scripts]
ensemble = "ensemble.cli:app"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "ruff>=0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ensemble"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src = ["src"]
line-length = 100
